+++ src/Lexer.php
@@ -8,11 +8,11 @@
  * Depends on context to extract lexemes.
  */
 
-namespace SqlParser;
+namespace PhpMyAdmin\SqlParser;
 
 require_once 'common.php';
 
-use SqlParser\Exceptions\LexerException;
+use PhpMyAdmin\SqlParser\Exceptions\LexerException;
 
 if (!defined('USE_UTF_STRINGS')) {
     // NOTE: In previous versions of PHP (5.5 and older) the default