+++ src/Utils/Formatter.php
@@ -4,12 +4,12 @@
  * Utilities that are used for formatting queries.
  */
 
-namespace SqlParser\Utils;
+namespace PhpMyAdmin\SqlParser\Utils;
 
-use SqlParser\Lexer;
-use SqlParser\Parser;
-use SqlParser\Token;
-use SqlParser\TokensList;
+use PhpMyAdmin\SqlParser\Lexer;
+use PhpMyAdmin\SqlParser\Parser;
+use PhpMyAdmin\SqlParser\Token;
+use PhpMyAdmin\SqlParser\TokensList;
 
 /**
  * Utilities that are used for formatting queries.