+++ tests/Components/FunctionCallTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace SqlParser\Tests\Components;
+namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use SqlParser\Components\ArrayObj;
-use SqlParser\Components\FunctionCall;
-use SqlParser\Tests\TestCase;
+use PhpMyAdmin\SqlParser\Components\ArrayObj;
+use PhpMyAdmin\SqlParser\Components\FunctionCall;
+use PhpMyAdmin\SqlParser\Tests\TestCase;
 
 class FunctionCallTest extends TestCase
 {