+++ tests/Lexer/IsMethodsTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace SqlParser\Tests\Lexer;
+namespace PhpMyAdmin\SqlParser\Tests\Lexer;
 
-use SqlParser\Context;
-use SqlParser\Token;
-use SqlParser\Tests\TestCase;
+use PhpMyAdmin\SqlParser\Context;
+use PhpMyAdmin\SqlParser\Token;
+use PhpMyAdmin\SqlParser\Tests\TestCase;
 
 class IsMethodsTest extends TestCase
 {